feat(kernels): extract qwen3 kernel crate
