+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "third_party/flashinfer"]
-	path = third_party/flashinfer
+[submodule "crates/pegainfer-kernels/third_party/flashinfer"]
+	path = crates/pegainfer-kernels/third_party/flashinfer
 	url = https://github.com/flashinfer-ai/flashinfer
