+++ Cargo.toml
@@ -5,7 +5,11 @@ default-run = "pegainfer"
 version = "0.1.0"
 edition = "2024"
 
+[workspace]
+members = ["crates/pegainfer-kernels"]
+
 [dependencies]
+pegainfer-kernels = { path = "crates/pegainfer-kernels" }
 safetensors = "0.7"
 anyhow = "1.0"
 ndarray = "0.16"
@@ -44,9 +48,6 @@ zeromq = { git = "https://github.com/BugenZhao/zmq.rs", rev = "dde4ee5c48846b0cc
 
 [features]
 reference = []
-
-[build-dependencies]
-cc = "1.0"
 
 [dev-dependencies]
 criterion = { version = "0.8", features = ["html_reports"] }
