+++ bench_snapshots/rtx-5090/qwen3-4b.json
@@ -1,0 +1,84 @@
+{
+  "commit": "3448f87",
+  "date": "2026-05-03",
+  "model": "Qwen3-4B",
+  "gpu": "NVIDIA GeForce RTX 5090",
+  "prefill_heavy": {
+    "prompt_len": 10000,
+    "output_len": 1,
+    "metrics": {
+      "ttft_ms": {
+        "avg_ms": 501.71845,
+        "p50_ms": 501.93214600000005,
+        "p95_ms": 503.511198,
+        "p99_ms": 503.749592,
+        "max_ms": 503.749592,
+        "samples": 20
+      },
+      "first_decode_step_ms": null,
+      "steady_tpot_ms": null,
+      "e2e_ms": {
+        "avg_ms": 501.718507,
+        "p50_ms": 501.932229,
+        "p95_ms": 503.511238,
+        "p99_ms": 503.74962700000003,
+        "max_ms": 503.74962700000003,
+        "samples": 20
+      },
+      "generated_tokens": {
+        "min": 1,
+        "max": 1,
+        "avg": 1.0,
+        "samples": 20
+      },
+      "request_tok_s": 1.9931495164107431,
+      "decode_tok_s": null
+    }
+  },
+  "decode_heavy": {
+    "prompt_len": 1024,
+    "output_len": 256,
+    "metrics": {
+      "ttft_ms": {
+        "avg_ms": 43.963086,
+        "p50_ms": 43.980075,
+        "p95_ms": 44.021534,
+        "p99_ms": 44.046838,
+        "max_ms": 44.046838,
+        "samples": 20
+      },
+      "first_decode_step_ms": {
+        "avg_ms": 7.373056999999999,
+        "p50_ms": 7.3730139999999995,
+        "p95_ms": 7.381652,
+        "p99_ms": 7.384898,
+        "max_ms": 7.384898,
+        "samples": 20
+      },
+      "steady_tpot_ms": {
+        "avg_ms": 7.398491,
+        "p50_ms": 7.395912,
+        "p95_ms": 7.456175,
+        "p99_ms": 7.462056,
+        "max_ms": 7.63168,
+        "samples": 5080
+      },
+      "e2e_ms": {
+        "avg_ms": 1930.5532210000001,
+        "p50_ms": 1930.6517190000002,
+        "p95_ms": 1930.9557260000001,
+        "p99_ms": 1931.115229,
+        "max_ms": 1931.115229,
+        "samples": 20
+      },
+      "generated_tokens": {
+        "min": 256,
+        "max": 256,
+        "avg": 256.0,
+        "samples": 20
+      },
+      "request_tok_s": 132.60447688026656,
+      "decode_tok_s": 135.16450174336637
+    }
+  }
+}
