+++ crates/pegainfer-kernels/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+name = "pegainfer-kernels"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1.0"
+half = { version = "2.4", features = ["num-traits"] }
+cudarc = { version = "0.19.4", features = [
+    "cuda-version-from-build-system",
+    "cublas",
+    "cuda-12080",
+    "f16",
+    "nccl",
+] }
+
+[build-dependencies]
+cc = "1.0"
+
+[lints.clippy]
+pedantic = { level = "warn", priority = -2 }
+cast_lossless = "allow"
+cast_possible_truncation = "allow"
+cast_possible_wrap = "allow"
+cast_precision_loss = "allow"
+cast_sign_loss = "allow"
+collapsible_else_if = "allow"
+collapsible_if = "allow"
+missing_errors_doc = "allow"
+missing_panics_doc = "allow"
+module_name_repetitions = "allow"
+must_use_candidate = "allow"
+similar_names = "allow"
+too_many_arguments = "allow"
+too_many_lines = "allow"
+uninlined_format_args = "allow"
+upper_case_acronyms = "allow"
+doc_markdown = "allow"
+redundant_clone = "warn"
+unused_peekable = "warn"
+dbg_macro = "warn"
+exit = "warn"
+get_unwrap = "warn"
+print_stdout = "allow"
+print_stderr = "allow"
+rc_buffer = "warn"
+rc_mutex = "warn"
+rest_pat_in_fully_bound_structs = "warn"
+
+[lints.rust]
+unsafe_op_in_unsafe_fn = "warn"
+unreachable_pub = "warn"
