+++ crates/pegainfer-kernels/csrc/argmax.cu
