+++ crates/pegainfer-kernels/csrc/common.cuh
