+++ crates/pegainfer-kernels/csrc/conv1d.cu
