+++ crates/pegainfer-kernels/csrc/flashinfer_sampling.cu
