+++ crates/pegainfer-kernels/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod ffi;
+pub mod ops;
+pub mod paged_kv;
+pub mod tensor;
