+++ crates/pegainfer-kernels/src/ops/embedding.rs
