+++ crates/pegainfer-kernels/src/paged_kv.rs
@@ -1,0 +1,34 @@
+/// Page-first geometry used by paged-KV kernels.
+///
+/// This is kernel-facing shape metadata only. Pool allocation, page ownership,
+/// and request state live in the root runtime crate.
+#[derive(Clone, Copy, Debug)]
+pub struct PagedKvLayout {
+    pub page_size: usize,
+    pub num_layers: usize,
+    pub num_kv_heads: usize,
+    pub head_dim: usize,
+    /// Elements in one K (or V) block: page_size x num_kv_heads x head_dim.
+    pub kv_block_len: usize,
+    /// Elements between layers within a page: 2 x kv_block_len (K then V).
+    pub layer_stride: usize,
+    /// Elements per page (all layers): num_layers x layer_stride.
+    pub page_stride: usize,
+}
+
+impl PagedKvLayout {
+    pub fn new(num_layers: usize, num_kv_heads: usize, head_dim: usize, page_size: usize) -> Self {
+        let kv_block_len = page_size * num_kv_heads * head_dim;
+        let layer_stride = 2 * kv_block_len;
+        let page_stride = num_layers * layer_stride;
+        Self {
+            page_size,
+            num_layers,
+            num_kv_heads,
+            head_dim,
+            kv_block_len,
+            layer_stride,
+            page_stride,
+        }
+    }
+}
