+++ crates/pegainfer-kernels/tools/triton/gen_triton_aot.py
