+++ docs/resources/5090.md
@@ -1,0 +1,155 @@
+# 5090 Development Guide
+
+**Status**: Complete
+**TL;DR**: 5090 dev box workflow: repo at `/root/develop/xingming/pegainfer`, Qwen3-4B at `/data/Qwen3-4B`, venv at `/root/develop/xingming/pegainfer/.venv`, use `bash -ic` to load proxy settings from `.bashrc`, use `rsync` rather than `scp`, and validate the vLLM frontend through `/v1/models`, `/v1/completions`, streaming, and `/v1/chat/completions`. Tokenization is owned by `vllm-text`; pegainfer should not grow another tokenizer module.
+
+## Machine
+
+- SSH alias: `5090`
+- Repo path: `/root/develop/xingming/pegainfer`
+- Model path: `/data/Qwen3-4B`
+- Python venv: `/root/develop/xingming/pegainfer/.venv`
+- GPU target: RTX 5090, `PEGAINFER_CUDA_SM=120`
+- Proxy: `.bashrc` exports `http_proxy` and `https_proxy`; use `bash -ic` for remote commands that need network.
+
+## Sync
+
+Use `rsync`, not `scp`.
+
+Verified manifest sync from the Mac checkout:
+
+```bash
+rsync -av Cargo.toml rust-toolchain.toml 5090:/root/develop/xingming/pegainfer/
+```
+
+Verified lockfile pull after remote Cargo resolution:
+
+```bash
+rsync -av 5090:/root/develop/xingming/pegainfer/Cargo.lock /Users/mac/code/pegainfer/Cargo.lock
+```
+
+## Temporary Validation Worktree
+
+For the Qwen3 kernels-crate split, the main checkout at `/root/develop/xingming/pegainfer` already had unrelated local changes, so validation used a disposable rsync worktree at `/root/develop/xingming/pegainfer-kernels-crate-build`.
+
+That worktree copied FlashInfer from the initialized main checkout into `crates/pegainfer-kernels/third_party/flashinfer`, then ran the 5090 release build, `cargo test --release --no-run`, Qwen3-4B e2e, and `bench_serving snapshot`. The generated 5090 snapshot was pulled back to the Mac checkout as `bench_snapshots/rtx-5090/qwen3-4b.json`.
+
+## Setup
+
+Initialize submodules after a fresh clone; FlashInfer headers are required by `build.rs`.
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && git submodule update --init --recursive"'
+```
+
+Create the venv if missing and install the packages needed for build-time Triton AOT and Hugging Face downloads.
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && if [ ! -d .venv ]; then uv venv .venv; fi && uv pip install --python .venv/bin/python triton huggingface_hub"'
+```
+
+Download Qwen3-4B to `/data`.
+
+```bash
+ssh 5090 'bash -ic '\''cd /root/develop/xingming/pegainfer && .venv/bin/python -c "from huggingface_hub import snapshot_download; snapshot_download(repo_id=\"Qwen/Qwen3-4B\", local_dir=\"/data/Qwen3-4B\", allow_patterns=[\"*.safetensors\", \"*.json\", \"*.model\", \"*.txt\", \"tokenizer.*\", \"vocab.*\", \"merges.txt\"]); print(\"downloaded /data/Qwen3-4B\")"'\'''
+```
+
+Current downloaded revision:
+
+- `Qwen/Qwen3-4B`
+- commit `1cfa9a7208912126459214e8b04321603b3df60c`
+- model directory size: `7.6G`
+
+## Build
+
+Always use release mode and set the 5090 SM target explicitly.
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && PEGAINFER_CUDA_SM=120 cargo build --release"'
+```
+
+This completed successfully after submodules were initialized.
+
+## Validation
+
+The Qwen3-4B golden e2e command:
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && PEGAINFER_CUDA_SM=120 PEGAINFER_TEST_MODEL_PATH=/data/Qwen3-4B cargo test --release --test e2e -- --nocapture"'
+```
+
+Current result after regenerating `test_data/Qwen3-4B.json`: pass.
+
+Compile all test targets without running model-heavy cases:
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && PEGAINFER_CUDA_SM=120 cargo test --release --no-run"'
+```
+
+Current result after the vLLM tokenizer cleanup: pass.
+
+Regenerate Qwen3-4B golden data:
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && PEGAINFER_CUDA_SM=120 PEGAINFER_TEST_MODEL_PATH=/data/Qwen3-4B cargo test --release --test regen_test_data regen_test_data -- --ignored --exact --nocapture"'
+```
+
+Use `--exact`; otherwise the substring filter also matches `regen_test_data_qwen35`.
+
+Pull the regenerated file back:
+
+```bash
+rsync -av 5090:/root/develop/xingming/pegainfer/test_data/Qwen3-4B.json /Users/mac/code/pegainfer/test_data/Qwen3-4B.json
+```
+
+HTTP smoke command used for a live server:
+
+```bash
+ssh 5090 'bash -ic "cd /root/develop/xingming/pegainfer && PEGAINFER_CUDA_SM=120 RUST_LOG=info cargo run --release -- --model-path /data/Qwen3-4B --port 18080"'
+```
+
+Verified request:
+
+```bash
+ssh 5090 "bash -ic 'curl -sS --max-time 10 http://127.0.0.1:18080/v1/completions -H \"Content-Type: application/json\" -d \"{\\\"prompt\\\":\\\"Hello\\\",\\\"max_tokens\\\":8}\"'"
+```
+
+With the vLLM frontend replacement, the served model ID is currently the local model path (`/data/Qwen3-4B`).
+
+Models endpoint:
+
+```bash
+ssh 5090 "bash -ic 'curl -sS --max-time 10 http://127.0.0.1:18080/v1/models'"
+```
+
+Non-streaming completions smoke:
+
+```bash
+ssh 5090 "bash -ic 'curl -sS --max-time 20 http://127.0.0.1:18080/v1/completions -H \"Content-Type: application/json\" -d \"{\\\"model\\\":\\\"/data/Qwen3-4B\\\",\\\"prompt\\\":\\\"What is 2 + 2?\\\",\\\"max_tokens\\\":8,\\\"temperature\\\":0,\\\"add_special_tokens\\\":false,\\\"return_token_ids\\\":true}\"'"
+```
+
+Observed output begins with ` Also, can you explain how to`.
+
+Streaming completions smoke:
+
+```bash
+ssh 5090 "bash -ic 'curl -sS --max-time 20 -N http://127.0.0.1:18080/v1/completions -H \"Content-Type: application/json\" -d \"{\\\"model\\\":\\\"/data/Qwen3-4B\\\",\\\"prompt\\\":\\\"What is 2 + 2?\\\",\\\"max_tokens\\\":6,\\\"temperature\\\":0,\\\"add_special_tokens\\\":false,\\\"stream\\\":true}\" | head -20'"
+```
+
+Chat completions smoke:
+
+```bash
+ssh 5090 "bash -ic 'curl -sS --max-time 20 http://127.0.0.1:18080/v1/chat/completions -H \"Content-Type: application/json\" -d \"{\\\"model\\\":\\\"/data/Qwen3-4B\\\",\\\"messages\\\":[{\\\"role\\\":\\\"user\\\",\\\"content\\\":\\\"What is 2 + 2?\\\"}],\\\"max_tokens\\\":8,\\\"temperature\\\":0}\"'"
+```
+
+After smoke tests, confirm the server was stopped:
+
+```bash
+ssh 5090 'ss -ltnp | grep -E ":18080|:8000" || true'
+```
+
+## Frontend Boundary
+
+The local process starts `vllm-server`; `vllm-text` owns tokenizer loading, prompt encoding, streaming detokenization, chat templates, and reasoning parsing. pegainfer owns model loading, scheduling, CUDA execution, sampling, and token-id output over the local engine-core bridge.
+
+Do not reintroduce a pegainfer tokenizer wrapper or local OpenAI HTTP route layer unless `vllm-frontend-rs` cannot provide the behavior.
